Fix type mismatch in typed_size.

+++ src/main/cpp/loader/omicsds_schema.h
@@ -326,7 +326,7 @@ struct OmicsFieldData {
     return ((T*)data.data())[idx];
   }
   template<class T>
-  int typed_size() const {
+  size_t typed_size() const {
     return data.size() / sizeof(T);
   }
 };
